remove piGeometryWrite from this commit

+++ utilities/piWrite.m
@@ -193,7 +193,9 @@ if isequal(thisR.get('optics type'),'lens')
             % also copy the file over to local/thisScene/lens
             lensDir = fullfile(thisR.get('output dir'),'lens');
             if ~exist(lensDir,'dir'), mkdir(lensDir);end
-            copyfile(lensFile,fullfile(lensDir,[lensName, ext]));
+            if ~exist(fullfile(lensDir,[lensName, ext]),'file')
+                copyfile(lensFile,fullfile(lensDir,[lensName, ext]));
+            end
         end
 
     end
@@ -284,7 +286,7 @@ outputDir  = thisR.get('output dir');
 
 % We check for the overwrite here and we make sure there is also an input
 % directory to copy from.
-if ~isempty(inputDir)
+if ~isempty(inputDir) && ~strcmpi(inputDir,outputDir)
     sources = dir(inputDir);
     status  = true;
     for i = 1:length(sources)
